API: Expose a C++ API, that the C API now wraps
Most of the LiteCore API ‘class’ types (C4Database, etc.) are now bona
fide C++ classes whose API pretty much mirrors the C API. And the
C API is now simply implemented as trivial wrappers around these
methods.

+++ LiteCore/Storage/SQLiteKeyStore.cc
@@ -314,7 +314,7 @@ namespace litecore {
         stmt->bindNoCopy(VersionParam, rec.version.buf, (int)rec.version.size);
         stmt->bindNoCopy(BodyParam,    rec.body.buf, (int)rec.body.size);
         stmt->bindNoCopy(ExtraParam,   rec.extra.buf, (int)rec.extra.size);
-        stmt->bind      (FlagsParam,   rawFlags);
+        stmt->bind      (FlagsParam,   (long long)rawFlags);
         stmt->bindNoCopy(KeyParam,     (const char*)rec.key.buf, (int)rec.key.size);
         stmt->bind      (SequenceParam,(long long)seq);
